reshuffles code to clear up the router and move Barcode specific code into it's own module
