+++ routes/barcode.js
@@ -1,10 +1,8 @@
 const express = require('express');
-const download = require('image-downloader');
-const graphicsmagick = require("gm");
 const fs = require("fs");
-const twitter = require('twitter');
 const router = express.Router();
 const article = require('../modules/Article');
+const barcode = require('../modules/Barcode');
 
 router.get('/', async (req, res, next) => {
   const width = (req.query.width) ? req.query.width : 1024;
@@ -16,170 +14,46 @@ router.get('/', async (req, res, next) => {
   const share = (req.query.share) ? req.query.share : '';
 
   try {
-    const images = await article.getImagesFromDateRange(dateFrom, dateTo);
-    const config = createConfig(orientation, fit, images.length, width, height);
-    const updatedImages = createImages(images, fit, config.width, config.height);
-    const downloadPath = process.env.DOWNLOAD_PATH;
-    const outputPath = `${downloadPath}_result/output.jpg`;
-    const imagePromises = [];
+    const downloadPath = `${process.env.DOWNLOAD_FOLDER}`;
+    const resultPath = `${process.env.RESULT_FOLDER}`;
+    const outputPath = `${resultPath}/output.jpg`;
     
-    updatedImages.forEach((image, i) => {
-      const newPromise = new Promise(function(resolve, reject) {
-        const options = {
-          url: image,
-          dest: `${downloadPath}${pad((i + 1), 5, '0')}.jpg`
-        };
-        
-        download.image(options)
-          .then(({ filename, image }) => {
-            if(orientation === 'h'){
-              graphicsmagick(filename)
-                .resize(config.width, config.span, "!")
-                .write(filename, function (err) {
-                  if (err) {
-                    console.log(err);
-                    res.json({ error: err });
-                  }
-                  resolve();
-                });
-            } else {
-              graphicsmagick(filename)
-                .resize(config.span, config.height, "!")
-                .write(filename, function (err) {
-                  if (err){
-                    console.log(err);
-                    res.json({ error: err });
-                  }
-                  resolve();
-                });
-            }
-          })
-          .catch((err) => {
-            console.error(err);
-            res.json({ error: err });
-            reject();
-          });
-      });
+    if (!fs.existsSync(downloadPath) || !fs.existsSync(resultPath)) {
+      res.json({ error: "Download folders not found" });
+    }
 
-      imagePromises.push(newPromise);
-    });
+    const images = await article.getImagesFromDateRange(dateFrom, dateTo);
+    const config = barcode.createConfig(orientation, fit, images.length, width, height);
+    const updatedImages = barcode.createImagePaths(images, fit, config.width, config.height);
+    const imagePromises = barcode.getImages(config, updatedImages, orientation, downloadPath);
 
     Promise.all(imagePromises)
       .then(function(values) {
-        let promise = new Promise(function(resolve, reject) {
-          let renderGm = graphicsmagick();
-
-          for(let i = 0; i < imagePromises.length; i++){
-            const name = pad((i + 1), 5, '0');
-            const pos = (i * config.span);
-
-            if(orientation === 'h'){
-              renderGm.in('-page', `+0+${pos}`)
-                .in(`${downloadPath}${name}.jpg`);
-            } else {
-              renderGm.in('-page', `+${pos}+0`)
-                .in(`${downloadPath}${name}.jpg`);
-            }
-          }
-
-          renderGm.mosaic()
-            .write(outputPath, function (err) {
-                if (err) console.log(err);
-                resolve();
-            });
-        });
-        
-        promise
-          .then(function(value) {
-            if(share === 'twitter'){
-              postTwitter('I am a tweet', outputPath);
-            }
-
+        const finalImage = barcode.createStitchedImage(config, imagePromises, downloadPath, orientation, outputPath);
+        finalImage
+          .then(() => shareCheck(share, outputPath))
+          .then(() => {
             let img = fs.readFileSync(outputPath);
             res.writeHead(200, {'Content-Type': 'image/jpg' });
             res.end(img, 'binary');
           })
           .catch((err) => {
-            console.error(err);
+            res.json({ error: `finalImage: ${err}` });
           });
       })
       .catch((err) => {
-        console.error(err);
+        res.json({ error: `imagePromises: ${err}` });
       });
 	} catch (err) {
-		next(err);
+    next(err);
+    res.json({ error: `router: ${err}` });
 	}
 });
 
-function createConfig(orientation, fit, num, width, height){
-  const config = {
-    width: width,
-    height: height,
-    span: 0,
-    fit: fit
-  };
-
-  if(fit === 'fill'){
-    if(orientation === 'h'){
-      config.span = Math.ceil( (height/100) * (100 / num) );
-    } else {
-      config.span = Math.ceil( (width/100) * (100 / num) );
-    }
-  } else {
-    if(orientation === 'h'){
-      config.height = Math.ceil( (height/100) * (100 / num) );
-      config.span = config.height;
-    } else {
-      config.width = Math.ceil( (width/100) * (100 / num) );
-      config.span = config.width;
-    }
+function shareCheck(share, outputPath){
+  if(share === 'twitter'){
+    barcode.postTwitter('I am a tweet', outputPath);
   }
-  return config;
-}
-
-function createImages(images, fit, width, height){
-  if(fit === 'fill'){
-    return images.map(url => {
-      const parts = url.split('?');
-      return parts[0].concat(`?source=ftlabs-barcode&width=10&height=10&quality=highest&fit=${fit}`);
-    });
-  }
-  
-  return images.map(url => {
-    const parts = url.split('?');
-    return parts[0].concat(`?source=ftlabs-barcode&width=${width}&height=${height}&quality=highest&fit=${fit}`);
-  });
-}
-
-function postTwitter(message, mediaPath){
-  let client = new twitter({
-    consumer_key: process.env.TWITTER_CONSUMER_KEY,
-    consumer_secret: process.env.TWITTER_CONSUMER_SECRET,
-    access_token_key: process.env.TWITTER_ACCESS_TOKEN_KEY,
-    access_token_secret: process.env.TWITTER_ACCESS_TOKEN_SECRET
-  });
-
-  let data = fs.readFileSync(mediaPath);
-
-  client.post('media/upload', {media: data}, function(error, media, response) {
-    if (!error) {
-      let status = {
-        status: message,
-        media_ids: media.media_id_string
-      }
-      client.post('statuses/update', status, function(error, tweet, response) {
-        if (!error) {
-          console.log(tweet);
-        }
-      });
-    }
-  });
-}
-
-function pad(n, width, z) {
-  z = z || '0';
-  n = n + '';
-  return n.length >= width ? n : new Array(width - n.length + 1).join(z) + n;
 }
 
 module.exports = router;
